requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
werkzeug==2.3.7
flask-cors==3.0.10
python-dotenv
supabase-py
